+++ src/bot/botrulez/ping.rs
@@ -2,12 +2,8 @@ use async_trait::async_trait;
 use clap::Parser;
 
 use crate::api::Message;
-use crate::bot::command::{ClapCommand, Context};
+use crate::bot::command::{ClapCommand, Command, Context};
 use crate::conn;
-
-/// Trigger a short reply.
-#[derive(Parser)]
-pub struct Args {}
 
 pub struct Ping(pub String);
 
@@ -22,6 +18,29 @@ impl Default for Ping {
         Self::new("Pong!")
     }
 }
+
+#[async_trait]
+impl<B, E> Command<B, E> for Ping
+where
+    E: From<conn::Error>,
+{
+    async fn execute(
+        &self,
+        arg: &str,
+        msg: &Message,
+        ctx: &Context,
+        _bot: &mut B,
+    ) -> Result<(), E> {
+        if arg.trim().is_empty() {
+            ctx.reply(msg.id, &self.0).await?;
+        }
+        Ok(())
+    }
+}
+
+/// Trigger a short reply.
+#[derive(Parser)]
+pub struct Args {}
 
 #[async_trait]
 impl<B, E> ClapCommand<B, E> for Ping